% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sp_hum.R
\name{calc_sp_hum}
\alias{calc_sp_hum}
\title{Calculate specific humidity}
\usage{
calc_sp_hum(airt, airp, hum, method)
}
\arguments{
\item{airt}{vector; air temperature values (Celsius)}

\item{airp}{vector; surface air pressure values (Pa)}

\item{hum}{vector; humidity values (either relative humidity \% [0-100] or dewpoint temperature [degC] or wet bulb temperature). Method}

\item{method}{character; Determine humidity values supplied to `hum`, either relative humidity 'rhum' or dewpoint temperature 'dewt' or wet bulb temperature 'wet_bulb'}
}
\value{
vector of specific humidity in kg/kg
}
\description{
Calculates specific humidity using air temperature and relative humidity or dewpoint temperature or wet bulb temperature. Using formula from GOTM source code: https://github.com/gotm-model/code/blob/master/src/airsea/humidity.F90
}
\examples{
 met_file <- system.file('extdaairt/met_file.dat', package = 'GOTMr')
 met <- read.delim(met_file)
 sp_hum <- calc_sp_hum(airt = met$AirT, airp = (met$MSLP*100), hum = met$DewT, method = 'dewt')
 plot(sp_hum)
}
