% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_hypso.R
\name{conv_hypso}
\alias{conv_hypso}
\title{Convert hypsograph to rLakeAnalyzer format}
\usage{
conv_hypso(in_file, out_file, header = F, sep = "\\t")
}
\arguments{
\item{in_file}{value file; File path to GOTM hypsograph file}

\item{out_file}{value file; File path to write the new ".bth' file to}

\item{header}{logical; indicating whether the file contains the names of the variables as its first line. Defaults to FALSE.}

\item{sep}{character; field separator character. Values on each line of the file are separated by this character. Defaults to tab}
}
\value{
File with hypsograph formatted for rLakeAnalyzer
}
\description{
Convert hypsograph format for GOTM to the format for use with rLakeAnalyzer
}
