% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_init_prof.R
\name{create_init_prof}
\alias{create_init_prof}
\title{Create initial profile for GOTM}
\usage{
create_init_prof(obs_file, date, tprof_file, print = TRUE, ...)
}
\arguments{
\item{obs_file}{filepath; Path to observation file}

\item{date}{character; Date in "YYYY-mm-dd HH:MM:SS" format to extract the initial profile.}

\item{tprof_file}{filepath; For the new initial temperature profile file.}

\item{print}{logical; Prints the temperature profile to the con}

\item{...}{arguments to be passed to read.delim() for reading in observed file e.g "header <- TRUE, sep <- ','"}
}
\value{
Message stating if the file has been created
}
\description{
Extract and format the initial profile for GOTM from the observation file used in ACPy.
}
