% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsummary.R
\name{dbsummary}
\alias{dbsummary}
\title{Summarise parsac database}
\usage{
dbsummary(dbFile, parsacXML)
}
\arguments{
\item{dbFile}{database file; filepath to database file with parsac results}

\item{parsacXML}{XML file; filepath to the xml file used for parsac calibration}
}
\value{
dataframe containing summary information about the database
}
\description{
Summarises all information within a database
}
