% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_plots}
\alias{diag_plots}
\title{Produces model diagnostic plots}
\usage{
diag_plots(mod, obs, ggplot = T)
}
\arguments{
\item{mod}{dataframe; modelled values in long format.}

\item{obs}{dataframe; observed values in long format.}

\item{ggplot}{logical; plot in ggplot or base plot. Defaults to TRUE}
}
\value{
grob object which can be assigned and then saved using ggsave() function
}
\description{
Produces 6 plots to assess model performance.
}
