% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDepths.R
\name{extractDepths}
\alias{extractDepths}
\title{Extract values at a depth or series of depths}
\usage{
extractDepths(
  mod.val,
  mod.dep,
  ncdf = NULL,
  var = NULL,
  depths = NULL,
  method = "linear",
  print = T
)
}
\arguments{
\item{mod.val}{dataframe of modelled values loaded using load.3d}

\item{mod.dep}{dataframe of depths of corresponding modelled values loaded}

\item{ncdf}{filepath; Name of the netCDF file to extract variable}

\item{var}{character; Name of the variable to be extracted. Must match short names in the netcdf file. Use list_vars() to view variables}

\item{depths}{vector of depth(s) for which to extract the value at}

\item{method}{character determining method for extracting depths at interpolated temperatures. Either linear interpolation or cubic spline. Defaults to linear.}

\item{print}{logical; Print a progress bar to the console. Defaults to TRUE}
}
\value{
data frame in the long format the same as load.obs(). With modelled values at the selected depths.
}
\description{
Extracts a variable at a certain depth or series of depths from netCDF or from text output
}
