% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_1d.R
\name{get_1d}
\alias{get_1d}
\title{Extract 1-D variable from NetCDF file}
\usage{
get_1d(ncdf, var)
}
\arguments{
\item{ncdf}{filepath; Name of the netCDF file to extract variable}

\item{var}{character; Name of the variable to be extracted. Must match}
}
\value{
vector; 1D object
}
\description{
Extracts a selected 1-D parameter from the netCDF file e.g. 'lat', 'lon'.
}
