% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flux.R
\name{get_flux}
\alias{get_flux}
\title{Extract fluxes from NetCDF file}
\usage{
get_flux(ncdf)
}
\arguments{
\item{ncdf}{filepath; Name of the netCDF file to extract variable}
}
\value{
dataframe with the DateTime and heat fluxes in each columns.
}
\description{
Extracts fluxes from the netCDF output file; Sensible heat flux (Qe), latent heat flux (Qh), long-wave back radation (Qb), net surface heat flux (heat).
}
