% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_param}
\alias{get_param}
\title{Extract dataframe of all parameters from ACPy database}
\usage{
get_param(dbFile, acpyXML, run = NULL)
}
\arguments{
\item{dbFile}{database file; filepath to database file where calibration results are stored}

\item{acpyXML}{XML file; filepath to xml file used for ACPy calibration}

\item{run}{numeric; Run number to extract parameters from. If NULL extracts runs from the whole database. Defaults to NULL}
}
\value{
data
}
\description{
Extract all parameters from database file with xml file
}
