% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vari_z.R
\name{get_vari_z}
\alias{get_vari_z}
\title{Extract variables from NetCDF file for a certain depth}
\usage{
get_vari_z(ncdf, var, z, constant_z = F)
}
\arguments{
\item{ncdf}{filepath; Name of the netCDF file to extract variable}

\item{var}{character; Name of the variable to be extracted. Must match short name in netCDF file}

\item{z}{numeric; Depth at which the variable should be extracted. Must be negative.}

\item{constant_z;}{If set to TRUE, only the first row of the depth data is used. If FALSE, the depth data is checked for varying water levels. Defaults to FALSE.}
}
\value{
dataframe/data.table
}
\description{
Extracts a selected parameter from the netCDF file for a certain depth and formats it into a dataframe.
}
\examples{
\dontrun{
get_vari_z("output.nc", "temp", z = -1)
}
}
\author{
Jorrit Mesman
}
