% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yaml_value.R
\name{get_yaml_value}
\alias{get_yaml_value}
\title{Extract values frpm yaml file}
\usage{
get_yaml_value(file = "gotm.yaml", label, key)
}
\arguments{
\item{file}{filepath; to .yaml which you wish to edit}

\item{label}{string; which corresponds to section where the key is located}

\item{key}{string; name of key in which to input the value}

\item{value}{string; to be input into the key/value pair. Note boolean values must be input as 'true'/'false' as per the yaml format}

\item{out_file}{filepath; to write the output yaml file (optional); defaults to overwriting file if not specified}
}
\description{
Inputs values into yaml file by locating the label and key within the yaml file. Preserves comments (#) if present. NOTE: this does not use a yaml parser so if there are yaml formatting errors this function will not pick them up.
}
\examples{
yaml_file <- system.file('extdata/gotm.yaml', package = 'GOTMr')
get_yaml_value(file = yaml_file, label = 'airp', key = 'file')
}
\author{
Tadhg Moore
}
