% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_heatmap.R
\name{gotm_heatmap}
\alias{gotm_heatmap}
\title{Create heatmap from GOTM output}
\usage{
gotm_heatmap(
  data,
  depths,
  res = 100,
  title = "Heatmap",
  points = FALSE,
  rev.y.ax = FALSE,
  surface = FALSE,
  fixed.level = TRUE,
  facet = FALSE,
  ylim = NULL,
  zlab = "",
  ...
)
}
\arguments{
\item{data}{dataframe; values loaded with load.3d}

\item{depths}{dataframe; depths loaded with load.depths}

\item{res}{numeric; Resolution of the approximated surface along the x/y axis. Defaults to 100}

\item{title}{character; Title of the graph. Defaults to 'Heatmap'}

\item{points}{logical; Plot points of measured temp. Defaults to FALSE}

\item{rev.y.ax}{logical; Reverse y axis. Defaults to FALSE}

\item{surface}{logical; Plot surface level of the lake. Defaults to TRUE}

\item{fixed.level}{logical; If the lake level is a fixed elevation}

\item{facet}{logical; option to plot a facet grid of the year data. Defaults to FALSE}

\item{ylim}{vector; Limits for the y-axis. Defaults to range of depths in the data}

\item{zlab}{character; Label for the Z axis}
}
\value{
Plot of interpolated heatmap
}
\description{
This function runs creates an interpolated heatmap from GOTM text output. It loads in a
water temp file generated by GOTM. It requires the depth '_z.txt' file also
output from GOTM to assign the depths to each measurement.
The level of interpolation can be adjusted to increase the accuracy
of the plot.
}
