% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_prof.R
\name{init_prof}
\alias{init_prof}
\title{Create initial profile for GOTM}
\usage{
init_prof(obs_file, date, tprof_file)
}
\arguments{
\item{date}{character; Date in "YYYY-mm-dd HH:MM:SS" format to extract the initial profile.}

\item{tprof_file}{filepath; For the new initial temperature profile file.}

\item{obs.file}{filepath; Path to observation file}
}
\value{
Message stating if the file has been created
}
\description{
Extract and format the initial profile for GOTM from the observation file used in ACPy.
}
