% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_nml.R
\name{input_nml}
\alias{input_nml}
\title{Edit nml file}
\usage{
input_nml(val, nml, par, print = T)
}
\arguments{
\item{val}{value; value to input into .nml file}

\item{nml}{.nml file; for the parameters to be input into}

\item{par}{parameter for which the value will be replaced}

\item{print}{logical; prints changes to .nml field to the console. Defaults to TRUE.}
}
\value{
data
}
\description{
Input values and file names into the GOTM namelist files
}
