% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_xml.R
\name{input_xml}
\alias{input_xml}
\title{Edit XML file}
\usage{
input_xml(param, xmlFile, print = T)
}
\arguments{
\item{param}{dataframe; parameters with colnames as parameter names which correspond to the parameters in the xml file}

\item{xmlFile}{XML file; for the parameters to be input into}

\item{print}{logical; prints changes to .xml fiel to the console. Defaults to TRUE.}
}
\value{
data
}
\description{
Input values and file names into the xml file for scenario setup
}
