% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_dates.R
\name{key_dates}
\alias{key_dates}
\title{Locate stratification or ice on/off dates}
\usage{
key_dates(data, type = "strat", days = 7, Tmin = 0.1, tz = "UTC")
}
\arguments{
\item{data}{dataframe; Water temperature profile data in the long form as loaded in using load.obs}

\item{type}{character; Which dates to caclulate, either ice or stratification on/off. Defaults to 'strat'.}

\item{days}{numeric; no. of days of consecutive days before identifying a break. Defaults to 7.}

\item{Tmin}{numeric; Minimum temperature as a cutoff point for identifying breaks. Defaults to 0.1}

\item{tz}{Timezone string to be supplied to as.POSIXct. Defaults to 'UTC'. This often can be left to the default unless timezone support is specifically required.}
}
\value{
vector with dates of stratification or ice on/off
}
\description{
Calculate dates of stratification or ice on/off dates. Ice on/off calculation is based on Pierson et al. (2011)
formula when the upper sensor records a temperature of at least -0.1C below that of the bottom sensor.
Stratification is based on the opposite principle i.e when there is a difference of +0.1C between the top and bottom sensor.
}
