% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_vars.R
\name{list_vars}
\alias{list_vars}
\title{List variables within the NetCDF file}
\usage{
list_vars(ncdf, long = F)
}
\arguments{
\item{ncdf}{filepath; To the location of the NetCDF file to be viewed.}

\item{long}{logical; Print the longnames of the variables, if FALSE it will print the short names. Defaults to FALSE}
}
\value{
Prints text to the console with the names of the variables from the NetCDF file.
}
\description{
Prints the short or long names of the variables within the NetCDF file
}
