% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_depths.R
\name{load_depths}
\alias{load_depths}
\title{Load layers file}
\usage{
load_depths(z.file, tz = "UTC")
}
\arguments{
\item{z.file}{depths file; File path to GOTM output file containing the depths which corresponds to the value file}

\item{tz}{Timezone string to be supplied to as.POSIXct. Defaults to 'UTC'. This often can be left to the default unless timezone support is specifically required.}
}
\value{
Dataframe with depths corresponding to values loaded via load.3d
}
\description{
Load layers file as a dataframe for use in plotting heatmap of the corresponding 3-d variable
}
