% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_obs.R
\name{load_obs}
\alias{load_obs}
\title{Load GOTMobservation file}
\usage{
load_obs(obs.file, header = F, sep = "\\t", tz = "UTC")
}
\arguments{
\item{obs.file}{observation file; File path to GOTM observation file containing the time, depths and values of observations}

\item{header}{logical; file contains the names of the variables as its first line. Defaults to FALSE}

\item{sep}{character; the delimiter in the file. Defaults to tab.}

\item{tz}{Assigns timezone to the DateTime column. Defaults to 'UTC'}
}
\value{
Dataframe of observation data for GOTM
}
\description{
Load observations which are used for calibration, for plotting in a heatmap
}
