% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_heatmap.R
\name{long_heatmap}
\alias{long_heatmap}
\title{Create heatmap from observed files for acpy}
\usage{
long_heatmap(obs, title = "Heatmap", ylim = NULL, xlim = NULL,
  size = 1, zlab = "", ...)
}
\arguments{
\item{obs}{dataframe; observations loaded with load.obs}

\item{title}{character; Title of the graph. Defaults to 'Heatmap'}

\item{ylim}{vector; Limits for the y-axis. Defaults to range of depths in the data}

\item{var.depths}{logical; If there is varying depths in the data.}

\item{res}{numeric; Resolution of the approximated surface along the x/y axis. Defaults to 100}

\item{contour}{logical; Add contour lines to plot. Defaults to FALSE}

\item{facet}{logical; Apply a facet_wrap and split the plots into different years. Default = FALSE}

\item{ncol}{numeric; number of columns for the facet plot. Default 2}

\item{nrow}{numeric; number of rows for the facet plot. Default 2}

\item{interp}{logical; Interpolate along y and x axis. Only use for daily observed data.}

\item{points}{logical; Plot points of measured temp. Defaults to FALSE}

\item{rev.y.ax}{logical; Reverse y axis. Defaults to FALSE}
}
\value{
Plot of interpolated heatmap
}
\description{
This function runs creates an interpolated heatmap from ACPy input files.
The level of interpolation can be adjusted to increase the accuracy
of the plot.
}
