% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_lineplot.R
\name{long_lineplot}
\alias{long_lineplot}
\title{Create line plot from observed temperature in long format}
\usage{
long_lineplot(data, rev.depths = FALSE, ylab = "Temperature",
  main = "Lineplot", colourblind = TRUE)
}
\arguments{
\item{data}{dataframe; observations loaded with load.obs}

\item{rev.depths}{logical; Reverse the labelling of the depths. Defaults to FALSE}

\item{ylab}{character; Label of the y-axis. Defaults to 'Temperature'}

\item{main}{character; Title of the graph. Defaults to 'Lineplot'}

\item{colourblind}{logical; Use colourblind friendly colours. Defaults to TRUE}

\item{xlim}{vector; Limits for the x-axis. Defaults to range of values in the data}
}
\value{
Line plot of temperatures at different depths
}
\description{
Plots a line plot of the temperatures with the lines coloured for each depth.
}
