% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_tstep.R
\name{match_tstep}
\alias{match_tstep}
\title{Match the timestep of two dataframes}
\usage{
match_tstep(df1, df2)
}
\arguments{
\item{df1}{dataframe; Data frame with timestep to be matched with.}

\item{df2}{dataframe; Data frame with data to be matched to the timestep of df1}
}
\value{
Dataframe of observation data subset to the same timestep as modelled data
}
\description{
Match the timestep of two different time series dataframes so they have the same number of steps. Works with long and wide-from data.
}
