% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDepths.R
\name{normDepths}
\alias{normDepths}
\title{Normalize depths of modelled output}
\usage{
normDepths(mod.dep, print = T)
}
\arguments{
\item{mod.dep}{dataframe of modelled values loaded using load.3d}

\item{print}{logical; Print a progress bar to the console. Defaults to TRUE}
}
\value{
data frame of normalized modelled depths
}
\description{
For modelled data where there is a varying lake level, this normalises the depths so the surface is the same and assigns depths to the depths lower relative to the top depth. Allows for comparison with observed data from load.obs
}
