% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE.R
\name{NSE}
\alias{NSE}
\title{Nash-Sutcliffe Efficiency}
\usage{
NSE(sim, obs, ...)
}
\arguments{
\item{sim}{numeric 'data.frame', 'matrix' or 'vector' with simulated values}

\item{obs}{numeric 'data.frame', 'matrix' or 'vector' with observed values}
}
\value{
data frame of normalized modelled depths
}
\description{
Nash-Sutcliffe efficiencies (Nash and Sutcliffe, 1970) range from -Inf to 1.
An efficiency of 1 (NSE = 1) corresponds to a perfect match of modeled to the observed data.
An efficiency of 0 (NSE = 0) indicates that the model predictions are as accurate
as the mean of the observed data, whereas
an efficiency less than zero (-Inf < NSE < 0) occurs when the observed mean is a better predictor than the model.
Essentially, the closer the model efficiency is to 1, the more accurate the model is.
}
