% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calib.R
\name{plot_calib}
\alias{plot_calib}
\title{Plot results from the database}
\usage{
plot_calib(
  dbFile,
  acpyXML,
  run = NULL,
  plot.dim = TRUE,
  nrow = 2,
  ncol = 4,
  zoom = FALSE,
  z.fct = 2,
  update = F
)
}
\arguments{
\item{dbFile}{database file; filepath to the database file where calibration results are stored}

\item{acpyXML}{XML file; filepath to the xml file used for ACPy calibration}

\item{run}{numeric; Run number to extract parameters from. If NULL extracts best parameters from the whole database}

\item{plot.dim}{logical; Control the dimensions of the plot area using nrow and ncol}

\item{nrow}{numeric; Only used if plot.dim is set to TRUE. Designates how many rows are in the plot. Defaults to 2}

\item{ncol}{numeric; Only used if plot.dim is set to TRUE. Designates how many columns are in the plot. Defaults to 4}

\item{zoom}{logical; Applies a zoom factor to the plot. Defaults to FALSE}

\item{z.fct}{numeric; Multiplication factor on zoom to the top of the graph. Defaults to 2. Only used if zoom is set to TRUE.}

\item{update}{logical; Creates a plot that updates every ~30 seconds. It needs to be manually cancelled. Defaults to FALSE..}
}
\value{
Plot of calibration
}
\description{
Creat a dotty plot of calibration results
}
