% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ext.R
\name{plot_ext}
\alias{plot_ext}
\title{Plot light extinction parameters}
\usage{
plot_ext(file, header = F, sep = "\\t", start = NULL, stop = NULL,
  tz = "UTC", main = "", ylab = "", type = "l", pch = ".")
}
\arguments{
\item{file}{value file; File path to GOTM meteorlogical input file.}

\item{header}{logical; indicating whether the file contains the names of the variables as its first line. Defaults to FALSE.}

\item{sep}{character; field separator character. Values on each line of the file are separated by this character. Defaults to tab}

\item{start}{string; Start date of range of data to be plotted.}

\item{stop}{string; Stop date of range of data to be plotted.}

\item{tz}{Timezone string to be supplied to as.POSIXct. Defaults to 'UTC'. This often can be left to the default unless timezone support is specifically required.}

\item{main}{string; an overall title for the plot}

\item{ylab}{string; a title for the y axis}

\item{type}{string; what type of plot should be drawn. Uses graphics::plot names. Defaults to 'line'.}

\item{pch}{string; Point character to be plotted. Defaults to '.'.}
}
\value{
Dataframe with Date time and values in wide format
}
\description{
Plots e-folding depths of visible and non-visible light.
}
