% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flux.R
\name{plot_flux}
\alias{plot_flux}
\title{Plot fluxes from NetCDF file}
\usage{
plot_flux(ncdf, title = "Heat Fluxes")
}
\arguments{
\item{ncdf}{filepath; Name of the netCDF file to extract variable}

\item{title}{character; Title of the graph. Defaults to 'Heat Fluxes'.}
}
\value{
dataframe with the
}
\description{
Plots fluxes from the netCDF output file; Sensible heat flux (Qe), latent heat flux (Qh), long-wave back radation (Qb), net surface heat flux (heat).
}
