% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hypso.R
\name{plot_hypso}
\alias{plot_hypso}
\title{Plot hypsograph data}
\usage{
plot_hypso(hypsograph_file, header = F, sep = "\\t", title = "Hypsograph")
}
\arguments{
\item{hypsograph_file}{value file; File path to GOTM hypsograph input file.}

\item{header}{logical; indicating whether the file contains the names of the variables as its first line. Defaults to FALSE.}

\item{sep}{character; field separator character. Values on each line of the file are separated by this character. Defaults to tab}

\item{title}{character; Title for the graph. Defaults to "Hypsograph"}
}
\value{
Plot of hypsograph
}
\description{
Plots hysograph curve for the lake
}
