% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mass_flux.R
\name{plot_mass_flux}
\alias{plot_mass_flux}
\title{Plot mass fluxes from NetCDF file}
\usage{
plot_mass_flux(ncdf, title = "Mass Fluxes")
}
\arguments{
\item{ncdf}{filepath; Name of the netCDF file to extract variable}

\item{title}{character; Title of the graph. Defaults to 'Mass Fluxes'.}
}
\value{
dataframe with the
}
\description{
Plots mass fluxes from the netCDF output file; Precipitation (precip)and evaporation (evap). Units are in m/s.
}
