% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meteo.R
\name{plot_meteo}
\alias{plot_meteo}
\title{Plot Meteo data}
\usage{
plot_meteo(meteo_file, header = F, sep = "\\t", start = NULL,
  stop = NULL, tz = "UTC")
}
\arguments{
\item{meteo_file}{value file; File path to GOTM meteorlogical input file.}

\item{header}{logical; indicating whether the file contains the names of the variables as its first line. Defaults to FALSE.}

\item{sep}{character; field separator character. Values on each line of the file are separated by this character. Defaults to tab}

\item{start}{string; Start date of range of data to be plotted.}

\item{stop}{string; Stop date of range of data to be plotted.}

\item{tz}{Timezone string to be supplied to as.POSIXct. Defaults to 'UTC'. This often can be left to the default unless timezone support is specifically required.}
}
\value{
Dataframe with Date time and values in wide format
}
\description{
Plots meteorlogical variables used for input into GOTM
}
