% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_month_prof.R
\name{plot_month_prof}
\alias{plot_month_prof}
\title{Plot monthly average profiles}
\usage{
plot_month_prof(mod, obs, year = FALSE, points = FALSE)
}
\arguments{
\item{mod}{dataframe; Modelled values in the long format.}

\item{obs}{dataframe; Observed values in the long format.}

\item{year}{boolean; Facet by year. Defaults to FALSE}

\item{points}{boolean; Include points for observation depths. Defaults to FALSE}
}
\value{
ggplot of temperature profiles
}
\description{
Plot monthly temperature averaged profiles for four months
}
