% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vari.R
\name{plot_vari}
\alias{plot_vari}
\title{Plot any variable from netcdf file}
\usage{
plot_vari(file, var, ...)
}
\arguments{
\item{file}{file; File path to netcdf file}

\item{var}{character; Name of the variable to be extracted. Must match short name in netCDF file}

\item{...}{other arguments to be passed to long_heatmap}
}
\value{
ggplot object; filled point plot of water temperature
}
\description{
Extracts modelled water temperature from netcdf file and plots it using long_heatmap function
}
