% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wbal.R
\name{plot_wbal}
\alias{plot_wbal}
\title{Plot water balance from NetCDF file}
\usage{
plot_wbal(ncdf, title = "Water Balance")
}
\arguments{
\item{ncdf}{filepath; Name of the netCDF file to extract variable}

\item{title}{character; Title of the graph. Defaults to 'Water Balance}
}
\value{
dataframe with the
}
\description{
Plots water balance calculated by GOTM ('Qres') and inflows from the netCDF output file.
}
