% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bestparam.R
\name{read_bestparam}
\alias{read_bestparam}
\title{Extract dataframe of best parameters from DB}
\usage{
read_bestparam(dbFile, acpyXML, run = NULL)
}
\arguments{
\item{dbFile}{database file; filepath to database file where calibration results are stored}

\item{acpyXML}{XML file; filepath to xml file used for ACPy calibration}

\item{run}{numeric; Run number to extract parameters from. If NULL extracts best parameters from the whole database. Defaults to NULL}
}
\value{
data
}
\description{
Extract best set of calibrated parameters from database file with xml file
}
