% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAcpyParams.R
\name{readAcpyParams}
\alias{readAcpyParams}
\title{Read ACPy configuration file and extract ranges}
\usage{
readAcpyParams(acpyXML)
}
\arguments{
\item{acpyXML}{XML file used for ACPy calibration}
}
\value{
data frame with parameter names and ranges
}
\description{
Read ACPy xml configuration file and extracts parameters and ranges in a dataframe
}
