% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_timeseries.R
\name{scan_timeseries}
\alias{scan_timeseries}
\title{Extract start and stop dates from a GOTM input file}
\usage{
scan_timeseries(file, header = F, sep = "\\t")
}
\arguments{
\item{file}{file; File path to ACPy observation file containing the time, depths and values of observations}

\item{header}{logical; file contains the names of the variables as its first line. Defaults to FALSE}

\item{sep}{character; the delimiter in the file. Defaults to tab}
}
\value{
vector with start and stop dates as a character
}
\description{
Extracts start and stop dates from a GOTM input file
}
