% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAcpyDb.R
\name{setAcpyDb}
\alias{setAcpyDb}
\title{Set ACPy database}
\usage{
setAcpyDb(acpyXML, name)
}
\arguments{
\item{acpyXML}{XML file used for ACPy calibration}

\item{name}{character; name of file for database, must end with '.db'}
}
\value{
Message saying what the database has been converted from
}
\description{
Read ACPy xml configuration file and set name for the database to save the results.
}
