% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setmodDepths.R
\name{setmodDepths}
\alias{setmodDepths}
\title{Extract modelled temperature at observation depths}
\usage{
setmodDepths(mod.val, mod.dep, obs = NULL, depths, method = "linear",
  print = T)
}
\arguments{
\item{mod.val}{dataframe of modelled values loaded using load.3d}

\item{mod.dep}{dataframe of depths of corresponding modelled values loaded using load.depths and corrected using normDepths()}

\item{obs}{dataframe of observations loaded in using load_obs. Default = NULL}

\item{depths}{vector; Depths at which to extract observed temperature (negative). Used instead of extracting depths from obs dataframe.}

\item{method}{character determining method for extracting depths at interpolated temperatures. Either linear interpolation or cubic spline. Defaults to linear.}

\item{print}{logical; Print a progress bar to the console. Defaults to TRUE}
}
\value{
data frame in the long format the same as load.obs(). With modelled values at the same depths as observed values.
}
\description{
Interpolates linearly ( or cubic spline) between model temperatures and extracts modelled temperatures at the same depths as observed temperatures or at specified depths.
}
