% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRange.R
\name{setRange}
\alias{setRange}
\title{Set parameter ranges for ACPy}
\usage{
setRange(acpyXML, param, min, max)
}
\arguments{
\item{acpyXML}{XML file used for ACPy calibration}

\item{param}{character corresponding to the name of the parameter in the ACPy xml file}

\item{min}{Minimum value for the parameter range}

\item{max}{Maximum value for the parameter range}
}
\value{
Message stating the change in parameter range
}
\description{
Read ACPy xml configuration file and set ranges for the parameter
}
