% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setspinup.R
\name{setspinup}
\alias{setspinup}
\title{Set spin up year for ACPy}
\usage{
setspinup(acpyXML, years)
}
\arguments{
\item{acpyXML}{XML file used for ACPy calibration}

\item{years}{numeric; no. of spin up years for ACPy calibration}
}
\value{
Message stating what the file name change was
}
\description{
Read ACPy xml configuration file and set number of spin up years
}
