% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams_switch.R
\name{streams_switch}
\alias{streams_switch}
\title{Switch on/off streams in gotm.yaml file}
\usage{
streams_switch(file = "gotm.yaml", method)
}
\arguments{
\item{file}{filepath; to .yaml which you wish to edit}

\item{method;}{character; Either 'on' or 'off'}
}
\description{
Places '#' in front of the streams within the file or removes them, effectively switching off in/outflows.
}
\examples{
streams_switch(file = 'gotm.yaml', method = 'off')
}
\author{
Tadhg Moore
}
