% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stat.R
\name{sum_stat}
\alias{sum_stat}
\title{Caclulate summary stats of mod v obs}
\usage{
sum_stat(mod, obs, depth = F, na.rm = T, depth.range = NULL)
}
\arguments{
\item{mod}{vector or dataframe; Vector if no depth values otherwise Modelled values in the long format. i.e. same as observed in ACPy}

\item{obs}{vector or dataframe; Vector if no depth values otherwise Observed values in the long format loaded in using load.obs}

\item{depth}{logical; Depth values are included. Defaults to False}

\item{na.rm}{logical; Remove NA'values}

\item{depth.range}{vector; vector with a depth range to extract statistics at certain depths. Upper limit first e.g. c(-5,-10). Defaults to whole depth range.}
}
\value{
data frame of summary statistics
}
\description{
Calculate general summary statistics of modelled water temperature vs observed water temperature; Pearson's R, variance, covariance, bias, Nash-Sutcliffe Efficiency (NSE) and Root mean squared error (RMSE).
}
