% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stat_1var.R
\name{sum_stat_1var}
\alias{sum_stat_1var}
\title{Caclulate summary stats of one variable}
\usage{
sum_stat_1var(mod, obs, na.rm = T)
}
\arguments{
\item{mod}{vector; Modelled values}

\item{obs}{vector; Observed values}

\item{na.rm}{logical; Remove NA'values}
}
\value{
data frame of summary statistics
}
\description{
Calculate general summary statistics of one variable, modelled vs observed; Pearson's R, variance, covariance, bias, Nash-Sutcliffe Efficiency (NSE) and Root mean squared error (RMSE).
}
