% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_mld.R
\name{ts_mld}
\alias{ts_mld}
\title{Create a timeseries of mixed layer depth}
\usage{
ts_mld(wtr, slope = 0.1, seasonal = TRUE, daily = F,
  mixed.cutoff = 1, tz = "UTC")
}
\arguments{
\item{wtr}{dataframe of water temperature in the long form as loaded through load.obs or formatted using match.mod()}

\item{slope}{a numeric vector corresponding to the minimum slope}

\item{seasonal}{logical; indicates whether the seasonal thermocline should be returned. This is fed to thermo.depth, which is used as the starting point. The seasonal thermocline is defined as the deepest density gradient found in the profile. If FALSE, the depth of the maximum density gradient is used as the starting point.}

\item{daily}{logical; if TRUE will calculate the MLD at a daily time step. Otherwise it will calculate it at the same timestep as the input data. Defaults to FALSE.}

\item{mixed.cutoff}{A cutoff (C) where below this threshold, thermo.depth and meta.depths are not calculated (NaN is returned). Defaults to 1 deg C.}

\item{tz}{Timezone string to be supplied to as.POSIXct. Defaults to 'UTC'. This often can be left to the default unless timezone support is specifically required.}
}
\value{
data frame of calculated mixed layer depth.
}
\description{
Calculates mixed layer depth (MLD) at each time step or on a daily time step. Returns NaN for steps where there is no MLD.
}
