% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_stat.R
\name{ts_stat}
\alias{ts_stat}
\title{Calculate time-series over a depth dependent variable}
\usage{
ts_stat(mod, obs, depth.range = NULL, stat = "pearson", tz = "UTC")
}
\arguments{
\item{mod}{dataframe; Modelled values in the long format. i.e. same as observed in ACPy}

\item{obs}{dataframe; Observed values in the long format loaded in using load.obs}

\item{depth.range}{vector; vector with a depth range to extract RMSE at certain depths. Upper limit first e.g. c(-5,-10). Defaults to whole water column.}

\item{stat}{'character; which statistic to use. Options are 'RMSE', 'NSE','MAE' and 'pearson'}

\item{tz}{Timezone string to be supplied to as.POSIXct. Defaults to 'UTC'. This often can be left to the default unless timezone support is specifically required.}
}
\value{
data frame; time series of RMSE
}
\description{
Calculate a time series of a statistical measure of modelled versus observed water temperature.
}
