% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_init.R
\name{view_init}
\alias{view_init}
\title{View inital profile}
\usage{
view_init(file, start, header = FALSE, sep = "\\t")
}
\arguments{
\item{file}{value file; File path to GOTM initial temperature file.}

\item{start}{string; Start date of model initialization.}

\item{header}{logical; indicating whether the file contains the names of the variables as its first line. Defaults to FALSE.}

\item{sep}{character; field separator character. Values on each line of the file are separated by this character. Defaults to tab}
}
\description{
Prints initial temperature profile for the model
}
