% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_nc.R
\name{view_nc}
\alias{view_nc}
\title{View NetCDF file}
\usage{
view_nc(ncdf)
}
\arguments{
\item{ncdf}{filepath; To the location of the NetCDF file to be viewed.}
}
\value{
Prints text to the console with details from the NetCDF file.
}
\description{
Prints the data from a netCDF file for viewing.
}
