% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2long.R
\name{wide2long}
\alias{wide2long}
\title{Converts wide data to long data}
\usage{
wide2long(data, depths)
}
\arguments{
\item{data}{dataframe; values loaded with load.3d}

\item{depths}{dataframe; depths loaded with load.depths}
}
\value{
dataframe in the same format as the observation file
}
\description{
Convert a dataframe in wide format to long format.
}
