% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cc.R
\name{calc_cc}
\alias{calc_cc}
\title{Calculate cloud cover}
\usage{
calc_cc(date, airt, relh = NULL, dewt = NULL, swr, lat, long, elev,
  daily = F)
}
\arguments{
\item{date}{vector; Dates in as.POSixct class}

\item{airt}{vector; Air temperature values which correspond to the vector of dates}

\item{relh}{vector; Relative humidity values which correspond to the vector of dates}

\item{dewt}{vector; Dewpoint temperature values which correspond to the vector of dates. Used instead of relative humidity}

\item{swr}{vector; Short-wave radiation values which correspond to the vector of dates}

\item{lat}{numeric; Latitude position (in decimal)}

\item{long}{numeric; Longitude position (in decimal)}

\item{elev}{numeric; elevation in metres above sea level}

\item{daily}{logical; Is the data on a daily timestep. Defaults to FALSE}
}
\value{
vector of cloud cover values which correspond to the vector of dates supplied
}
\description{
Calculate cloud cover using latitude, air temperature, relative humidity (or dewpoint temperature) and short wave radiation using the calculations from Martin and McCutcheon (1999).
}
