% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_in_lwr.R
\name{calc_in_lwr}
\alias{calc_in_lwr}
\title{Calculate incoming Long-wave radiation}
\usage{
calc_in_lwr(cc, airt, relh = NULL, dewt = NULL)
}
\arguments{
\item{cc}{vector; cloud cover values (Fraction [0-1])}

\item{airt}{vector; air temperature values (Celsius)}

\item{relh}{vector; relative humidity values (\% [0-100]). Used to calculate dewpoint temperaure if that is not supplied. Defaults to NULL.}

\item{dewt}{vector; dewpoint temperature values (Celsius). Can be used instead of relative humidity, otherise it is calculated from relative humidity and air temperature. Defaults to NULL}
}
\value{
vector of incoming long-wave radiation values
}
\description{
Calculates incoming LWR using cloud cover, air temperature and relative humidity (or dewpoint temperature). Using formula from: http://www.seao2.info/_TMP/longwave.pdf
}
\examples{
 met_file <- system.file('extdata/met_file.dat', package = 'GOTMr')
 met <- read.delim(met_file)
 lwr_in <- calc_in_lwr(cc = met$CC, airt = met$AirT, dewt = met$DewT)
 plot(lwr_in)
}
