% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbsummary.R
\name{dbsummary}
\alias{dbsummary}
\title{Summarise ACPy database}
\usage{
dbsummary(dbFile, acpyXML)
}
\arguments{
\item{dbFile}{database file; filepath to database file with ACPy results}

\item{acpyXML}{XML file; filepath to the xml file used for ACPy calibration}
}
\value{
dataframe containing summary information about the database
}
\description{
Summarises all information within a database
}
