% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_plots}
\alias{diag_plots}
\title{Produces model diagnostic plots}
\usage{
diag_plots(mod, obs, size = 0.1, ggplot = TRUE, colourblind = FALSE,
  na.rm = FALSE)
}
\arguments{
\item{mod}{dataframe; modelled values in long format.}

\item{obs}{dataframe; observed values in long format.}

\item{size}{numeric; size of points in plot. Defaults to 0.1}

\item{ggplot}{logical; plot in ggplot or base plot. Defaults to TRUE}

\item{colourblind}{logical; Use colourblind friendly colours. Defaults to FALSE}

\item{na, rm}{boolean; a logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to FALSE}
}
\value{
grob object which can be assigned and then saved using ggsave() function
}
\description{
Produces 6 plots to assess model performance.
}
