% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flux.R
\name{get_flux}
\alias{get_flux}
\title{Extract fluxes from NetCDF file}
\usage{
get_flux(ncdf, agg.dly = FALSE)
}
\arguments{
\item{ncdf}{filepath; Name of the netCDF file to extract variable}

\item{agg.dly}{logical; Aggregate the fluxes to a daily timestep.}
}
\value{
dataframe with the DateTime and heat fluxes in each columns.
}
\description{
Extracts fluxes from the netCDF output file; Sensible heat flux (Qe), latent heat flux (Qh), long-wave back radation (Qb), incoming solar radiation (Qsw) and calculates the net sum of all these fluxes (Net).
Net = Qe + Qh + Qb + Qsw
Negative flux represents upward flux while positive indicated downward flux.
}
