% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_heatmap.R
\name{long_heatmap}
\alias{long_heatmap}
\title{Create heatmap from observed files for acpy}
\usage{
long_heatmap(obs, title = "Heatmap", ylim = NULL, xlim = NULL,
  size = 1, zlab = "", ...)
}
\arguments{
\item{obs}{dataframe; observations loaded with load.obs}

\item{title}{character; Title of the graph. Defaults to 'Heatmap'}

\item{ylim}{vector; Limits for the y-axis. Defaults to range of depths in the data}

\item{xlim}{vector; Limits for the x-axis. Defaults to range of values in the data}
}
\value{
Plot of interpolated heatmap
}
\description{
This function runs creates an interpolated heatmap from ACPy input files.
The level of interpolation can be adjusted to increase the accuracy
of the plot.
}
