% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_heatmap.R
\name{long_heatmap}
\alias{long_heatmap}
\title{Create heatmap from dataerved files for acpy}
\usage{
long_heatmap(
  data,
  title = "Heatmap",
  ylim = NULL,
  xlim = NULL,
  size = 1,
  zlab = "",
  ...
)
}
\arguments{
\item{data}{dataframe; dataervations loaded with load.data}

\item{title}{character; Title of the graph. Defaults to 'Heatmap'}

\item{ylim}{vector; Limits for the y-axis. Defaults to range of depths in the data}

\item{xlim}{vector; Limits for the x-axis. Defaults to range of values in the data}
}
\value{
Plot of interpolated heatmap
}
\description{
This function runs creates an interpolated heatmap from ACPy input files.
The level of interpolation can be adjusted to increase the accuracy
of the plot.
}
\examples{
sim_folder <- system.file('extdata', package = 'GOTMr')
run_gotm(sim_folder)
out <- file.path(sim_folder, 'output', 'output.nc')
wtemp <- get_vari(ncdf = out, var = 'temp')
z <- get_vari(ncdf = out, var = 'z')
df <- wide2long(data = wtemp, depths = z)
long_heatmap(data = df, title = 'Feeagh simulated temp', zlab = 'degC')
}
