% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_month_prof.R
\name{plot_month_prof}
\alias{plot_month_prof}
\title{Plot monthly average profiles}
\usage{
plot_month_prof(
  mod,
  obs,
  month = c("Jan", "Apr", "Jul", "Oct"),
  year = NULL,
  fixed.level = T,
  step = 2
)
}
\arguments{
\item{mod}{dataframe; Modelled values in the long format. i.e. same as observed in ACPy}

\item{obs}{dataframe; Observed values in the long format loaded in using load_obs}

\item{month}{vector; with string values for months for profiles in 3-letter format e.g. 'Jan'. Default = c('Jan','Apr', 'Jul','Oct')}

\item{year}{numeric; specifying year of profiles. Defaults to year of first measurement if not specified.}

\item{fixed.level}{logical; Temperature measurements are at fixed points throughout the timeseries. Defaults to TRUE}

\item{step}{numeric; Depth to average temperature measurements to. Only used if fixed.level = FALSE.}
}
\value{
plot of temperature profiles
}
\description{
Plot monthly temperature averaged profiles for four months
}
