% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vari.R
\name{plot_vari}
\alias{plot_vari}
\title{Plot any variable from netcdf file}
\usage{
plot_vari(ncdf, var, incl_time = TRUE, ...)
}
\arguments{
\item{ncdf}{filepath; to the netCDF file to extract variable}

\item{var}{character; Name of the variable to be extracted. Must match short name in netCDF file}

\item{incl_time}{boolean; Add time to the first column in the dataframe. Defaults to TRUE}

\item{...}{other arguments to be passed to long_heatmap}
}
\value{
ggplot object; filled point plot of water temperature
}
\description{
Extracts modelled water temperature from netcdf file and plots it using long_heatmap function
}
\examples{
sim_folder <- system.file('extdata', package = 'GOTMr')
run_gotm(sim_folder)
out <- file.path(sim_folder, 'output', 'output.nc')
plot_vari(ncdf = out, var = 'temp')
}
