% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wtemp.R
\name{plot_wtemp}
\alias{plot_wtemp}
\title{Plot wtemp from netcdf file}
\usage{
plot_wtemp(file, ...)
}
\arguments{
\item{file}{file; File path to netcdf file}

\item{...}{other arguments to be passed to long_heatmap}
}
\value{
ggplot object; filled point plot of water temperature
}
\description{
Extracts modelled water temperature from netcdf file and plots it using long_heatmap function
}
