% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2long.R
\name{wide2long}
\alias{wide2long}
\title{Converts wide data to long data}
\usage{
wide2long(data, depths)
}
\arguments{
\item{data}{dataframe; values loaded with get_vari()}

\item{depths}{dataframe; depths loaded with get_vari() or vector of depths corresponding to values in the datafra}
}
\value{
dataframe in the same format as the observation file
}
\description{
Convert a dataframe in wide format to long format.
}
\examples{
sim_folder <- system.file('extdata', package = 'GOTMr')
run_gotm(sim_folder)
out <- file.path(sim_folder, 'output', 'output.nc')
wtemp <- get_vari(ncdf = out, var = 'temp')
z <- get_vari(ncdf = out, var = 'z')
df <- wide2long(data = wtemp, depths = z)
head(df)
}
